fixed: error with data moved: api to init updated: version up to 0.0.2
